Ensure all other classes are registered before we initialize components
