+++ src/client/extensionActivation.ts
@@ -5,22 +5,16 @@
 
 import { CodeActionKind, debug, DebugConfigurationProvider, languages, OutputChannel, window } from 'vscode';
 
-import { registerTypes as activationRegisterTypes } from './activation/serviceRegistry';
 import {
     IExtensionActivationManager,
     IExtensionSingleActivationService,
     ILanguageServerExtension,
 } from './activation/types';
-import { registerTypes as appRegisterTypes } from './application/serviceRegistry';
 import { IApplicationDiagnostics } from './application/types';
 import { DebugService } from './common/application/debugService';
 import { IApplicationEnvironment, ICommandManager, IWorkspaceService } from './common/application/types';
-import { Commands, PYTHON, PYTHON_LANGUAGE, STANDARD_OUTPUT_CHANNEL, UseProposedApi } from './common/constants';
-import { registerTypes as installerRegisterTypes } from './common/installer/serviceRegistry';
+import { Commands, PYTHON, PYTHON_LANGUAGE, STANDARD_OUTPUT_CHANNEL } from './common/constants';
 import { traceError } from './common/logger';
-import { registerTypes as platformRegisterTypes } from './common/platform/serviceRegistry';
-import { IFileSystem } from './common/platform/types';
-import { registerTypes as processRegisterTypes } from './common/process/serviceRegistry';
 import { StartPage } from './common/startPage/startPage';
 import { IStartPage } from './common/startPage/types';
 import {
@@ -32,45 +26,36 @@ import {
 } from './common/types';
 import { OutputChannelNames } from './common/utils/localize';
 import { noop } from './common/utils/misc';
-import { registerTypes as variableRegisterTypes } from './common/variables/serviceRegistry';
 import { DebuggerTypeName } from './debugger/constants';
 import { DebugSessionEventDispatcher } from './debugger/extension/hooks/eventHandlerDispatcher';
 import { IDebugSessionEventHandlers } from './debugger/extension/hooks/types';
-import { registerTypes as debugConfigurationRegisterTypes } from './debugger/extension/serviceRegistry';
 import { IDebugConfigurationService, IDebuggerBanner } from './debugger/extension/types';
-import { registerTypes as formattersRegisterTypes } from './formatters/serviceRegistry';
 import {
     IComponentAdapter,
     IInterpreterLocatorProgressHandler,
     IInterpreterLocatorProgressService,
     IInterpreterService,
 } from './interpreter/contracts';
-import { registerTypes as interpretersRegisterTypes } from './interpreter/serviceRegistry';
 import { getLanguageConfiguration } from './language/languageConfiguration';
 import { LinterCommands } from './linters/linterCommands';
-import { registerTypes as lintersRegisterTypes } from './linters/serviceRegistry';
-import { addOutputChannelLogging, setLoggingLevel } from './logging';
+import { setLoggingLevel } from './logging';
 import { PythonCodeActionProvider } from './providers/codeActionProvider/pythonCodeActionProvider';
 import { PythonFormattingEditProvider } from './providers/formatProvider';
 import { ReplProvider } from './providers/replProvider';
-import { registerTypes as providersRegisterTypes } from './providers/serviceRegistry';
 import { activateSimplePythonRefactorProvider } from './providers/simpleRefactorProvider';
 import { TerminalProvider } from './providers/terminalProvider';
 import { ISortImportsEditingProvider } from './providers/types';
-import { setExtensionInstallTelemetryProperties } from './telemetry/extensionInstallTelemetry';
-import { registerTypes as tensorBoardRegisterTypes } from './tensorBoard/serviceRegistry';
-import { registerTypes as commonRegisterTerminalTypes } from './terminals/serviceRegistry';
 import { ICodeExecutionManager, ITerminalAutoActivation } from './terminals/types';
-import { TEST_OUTPUT_CHANNEL } from './testing/common/constants';
 import { ITestContextService } from './testing/common/types';
 import { ITestCodeNavigatorCommandHandler, ITestExplorerCommandHandler } from './testing/navigation/types';
-import { registerTypes as unitTestsRegisterTypes } from './testing/serviceRegistry';
 
 // components
 // import * as pythonEnvironments from './pythonEnvironments';
 
 import { ActivationResult, ExtensionState } from './components';
 import { Components } from './extensionInit';
+import { setExtensionInstallTelemetryProperties } from './telemetry/extensionInstallTelemetry';
+import { IFileSystem } from './common/platform/types';
 
 export async function activateComponents(
     // `ext` is passed to any extra activation funcs.
@@ -112,35 +97,9 @@ async function activateLegacy(ext: ExtensionState): Promise<ActivationResult> {
     const { context, legacyIOC } = ext;
     const { serviceManager, serviceContainer } = legacyIOC;
 
-    // register "services"
-
-    const standardOutputChannel = window.createOutputChannel(OutputChannelNames.python());
-    addOutputChannelLogging(standardOutputChannel);
-    const unitTestOutChannel = window.createOutputChannel(OutputChannelNames.pythonTest());
-    serviceManager.addSingletonInstance<OutputChannel>(IOutputChannel, standardOutputChannel, STANDARD_OUTPUT_CHANNEL);
-    serviceManager.addSingletonInstance<OutputChannel>(IOutputChannel, unitTestOutChannel, TEST_OUTPUT_CHANNEL);
-
-    // Core registrations (non-feature specific).
-    platformRegisterTypes(serviceManager);
-    processRegisterTypes(serviceManager);
-
     // We need to setup this property before any telemetry is sent
     const fs = serviceManager.get<IFileSystem>(IFileSystem);
     await setExtensionInstallTelemetryProperties(fs);
-
-    const applicationEnv = serviceManager.get<IApplicationEnvironment>(IApplicationEnvironment);
-    const enableProposedApi = applicationEnv.packageJson.enableProposedApi;
-    serviceManager.addSingletonInstance<boolean>(UseProposedApi, enableProposedApi);
-    // Feature specific registrations.
-    variableRegisterTypes(serviceManager);
-    unitTestsRegisterTypes(serviceManager);
-    lintersRegisterTypes(serviceManager);
-    interpretersRegisterTypes(serviceManager);
-    formattersRegisterTypes(serviceManager);
-    installerRegisterTypes(serviceManager);
-    commonRegisterTerminalTypes(serviceManager);
-    debugConfigurationRegisterTypes(serviceManager);
-    tensorBoardRegisterTypes(serviceManager);
 
     const configuration = serviceManager.get<IConfigurationService>(IConfigurationService);
     // We should start logging using the log level as soon as possible, so set it as soon as we can access the level.
@@ -150,13 +109,6 @@ async function activateLegacy(ext: ExtensionState): Promise<ActivationResult> {
 
     const abExperiments = serviceContainer.get<IExperimentsManager>(IExperimentsManager);
     await abExperiments.activate();
-
-    const languageServerType = configuration.getSettings().languageServer;
-
-    // Language feature registrations.
-    appRegisterTypes(serviceManager, languageServerType);
-    providersRegisterTypes(serviceManager);
-    activationRegisterTypes(serviceManager, languageServerType);
 
     // "initialize" "services"
 
@@ -178,6 +130,7 @@ async function activateLegacy(ext: ExtensionState): Promise<ActivationResult> {
     disposables.push(cmdManager.registerCommand(Commands.ViewOutput, () => outputChannel.show()));
     const startPage = serviceManager.get<StartPage>(IStartPage);
     cmdManager.registerCommand(Commands.OpenStartPage, () => startPage.open());
+    const applicationEnv = serviceManager.get<IApplicationEnvironment>(IApplicationEnvironment);
     cmdManager.executeCommand('setContext', 'python.vscode.channel', applicationEnv.channel).then(noop, noop);
 
     // Display progress of interpreter refreshes only after extension has activated.
@@ -198,6 +151,7 @@ async function activateLegacy(ext: ExtensionState): Promise<ActivationResult> {
     serviceManager.get<ITerminalAutoActivation>(ITerminalAutoActivation).register();
     const pythonSettings = configuration.getSettings();
 
+    const standardOutputChannel = window.createOutputChannel(OutputChannelNames.python());
     activateSimplePythonRefactorProvider(context, standardOutputChannel, serviceContainer);
 
     const sortImports = serviceContainer.get<ISortImportsEditingProvider>(ISortImportsEditingProvider);
